assignment3/requirements_gpu.txt
tensorflow_gpu>=1.12
sentencepiece
segtok
ipykernel
jupyter
tqdm
